First pass at methods to save results

+++ benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/package.scala
@@ -118,7 +118,7 @@ package object benchmarks {
                           mapping: Mapping,
                           query: NearestNeighborsQuery,
                           k: Int,
-                          recalls: Seq[Float],
+                          recalls: Seq[Double],
                           durations: Seq[Long])
 
 }
